Refactor Kafka Consumer initialization to accept group ID and update topic handling
