+++ services/preprocessor/config.py
@@ -1,6 +1,8 @@
 import os
 from typing import List
 
-KAFKA_BOOTSTRAP: List[str] = os.getenv("KAFKA_BOOTSTRAP", "kafka:9094").split(",")
+KAFKA_BOOTSTRAP: List[str] = os.getenv("KAFKA_BOOTSTRAP", "localhost:9094").split(",")
 KAFKA_TOPIC_ANTI = os.getenv("KAFKA_TOPIC_ANTI" ,"raw_tweets_antisemitic")
 KAFKA_TOPIC_NOT_ANTI = os.getenv("KAFKA_TOPIC_NOT_ANTI" ,"raw_tweets_not_antisemitic")
+
+GROUP_ID = os.getenv("GROUP_ID" ,"preprocessur_group")
